10/05/2022 16h53

+++ ft_strlen.c
@@ -6,11 +6,12 @@
 /*   By: adohou <marvin@42.fr>                      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/05/04 17:46:42 by adohou            #+#    #+#             */
-/*   Updated: 2022/05/08 20:58:55 by adohou           ###   ########.fr       */
+/*   Updated: 2022/05/10 16:16:42 by adohou           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
+#include "libft.h"
 
-int	ft_strlen(char *str)
+size_t	ft_strlen(char *str)
 {
 	int	i;
